move generator out of game
